add more granular type definitions for GraphQL types
